Hung up on the ligand
I'm glad we picked a tough ligand for the test. Currently, I can successfully get ligand parameters using antechamber if I use the SDF that was already in the folder, or a SDF exported from maestro after complex minimization. However, the resulting parameters have arbitrary atom names and don't match up with the PDB atoms when I do the final loadpdb command. If I try to use a PDB from any source, antechamber messes up with a carbon atom (one in a six-membered ring), and while the resulting parameters do match the atom names in the complex pdb, the topology is incomplete and causes tleap to crash.